Fix appendix section titles
The example style file has appendix titles formatted as "Appendix A.", but the current `.sty` file only outputs "A." by default. This can't be fixed by packages like `appendix` or `titlesec` due to the custom sectioning of the `.sty` file. This patch modifies the section display when the `\appendix` command is run so that the appendix titles are displayed correctly.